+++ .storybook/preview.js
@@ -2,8 +2,8 @@ import { addDecorator } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withInfo } from '@storybook/addon-info';
 import { withA11y } from '@storybook/addon-a11y';
-import setupI18n from '../demo/i18n/config';
-import '../demo/i18n/en';
+import setupI18n from './utils/i18n/config';
+import './utils/i18n/en';
 import './style/story-root.scss';
 import { withThemeSelector } from './theme-selector'
 import { addParameters } from '@storybook/react';
@@ -14,7 +14,14 @@ addParameters({
     panelPosition: 'bottom',
     showNav: true,
     showPanel: true,
-    theme: sageTheme
+    theme: sageTheme,
+    storySort: (a, b) => {
+      if (a[1].kind === 'Welcome' || b[1].kind === 'Welcome') {
+        return 1;
+      }
+      return a[1].id.localeCompare(b[1].id);
+      
+    }
   }
 });
 
@@ -28,3 +35,4 @@ addDecorator(withInfo({
 addDecorator(withA11y);
 addDecorator(withThemeSelector);
 
+window.STORYBOOK_GA_ID = 'UA-77028225-13';