+++ .storybook/welcome-page/common.style.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  margin: 0 auto;
+  max-width: 1024px;
+  padding: 25px 4%;
+  width: 92%;
+`;
+
+export const ContentWrapper = styled.div`
+  padding-top: 0px;
+  padding-bottom: 0px;
+  margin: 0 auto;
+  max-width: 1024px;
+  padding: 25px 4%;
+  width: 92%;
+`;