+++ cypress/locators/locators.js
@@ -12,6 +12,7 @@ export const HELP_ICON_PREVIEW = '[data-component="help"]';
 export const TOOLTIP_PREVIEW = '[data-component="tooltip"]';
 export const FIELD_HELP_PREVIEW = 'span[data-element="help"]';
 export const STORY_ROOT = '#story-root';
+export const DLS_ROOT = '#root';
 export const BACKGROUND_UI_LOCATOR = 'div[data-element="modal-background"]';
 export const LINK = '[data-component="link"]';
 export const ICON = '[data-component="icon"]';