+++ src/__experimental__/components/checkbox/checkbox.d.ts
@@ -9,7 +9,7 @@ interface CheckboxProps {
   inputWidth?: number | string;
   label?: string;
   labelAlign?: string;
-  labelWidth? : number | string;
+  labelWidth?: number | string;
   onChange?: (ev: React.ChangeEvent<HTMLElement>) => void;
   reverse?: boolean;
   size?: string;